feat: rework theming
