--- StatusSwift/Controls/ActionButtonControl.xaml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="StatusSwift.Controls.ActionButtonControl"
-             x:Name="actionButtonControl">
-    <Button
-        x:Name="ToggleButton"
-        Text="{Binding ButtonText, Source={x:Reference actionButtonControl}}"
-        Command="{Binding ButtonCommand, Source={x:Reference actionButtonControl}}"
-        HorizontalOptions="Center"
-        WidthRequest="220"
-        HeightRequest="60"
-        FontSize="18"
-        FontAttributes="Bold"
-        CornerRadius="12"
-        BackgroundColor="{Binding ButtonBackgroundColor, Source={x:Reference actionButtonControl}}"
-        Margin="0,10,0,0">
-        <Button.Shadow>
-            <Shadow Brush="{StaticResource Gray600}"
-                    Offset="0,2"
-                    Radius="4" />
-        </Button.Shadow>
-    </Button>
-</ContentView>
